Resume_Parse/Resume_Parse_Deploy/resume-matcher/requirements.txt
pdfplumber>=0.7.6
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
sqlalchemy>=1.4.0
numpy>=1.24.0
scikit-learn>=1.2.2
spacy>=3.6.0
typer>=0.9.0
python-dotenv>=1.0.0
tqdm>=4.65.0
# Optional if you want OpenAI embeddings:
openai>=1.0.0
